kargo gönderisi ekleme ve listeleme yapıldı. gerekli validasyonlar yapıldı. ek olarak kargo düzenleme yapıldı. türkçe validasyon ayarlandı.

+++ database/migrations/2024_03_19_155450_create_cargos_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('user_surname');
             $table->string('product_name');
             $table->string('product_content');
-            $table->unsignedBigInteger('product_kilogram');
+            $table->double('product_kilogram');
             $table->text('product_description');
             $table->boolean('status')->default(1);
             $table->timestamps();
